Move loop control device check into squashfs sanity check. Only check for loop-control if squashfs.FsType() is "squashfs", so it's skipped e.g. with lxd containers.
